DF/09: add possibility to specify timestamps time zone (set to UTC).
Previously 'naive' datetime objects were used, meaning in fact that we
treat `timestamp` field as timestamp in the local time zone. It has some
effect only when `now()` is used as a right limit of the integration
interval.

It is said (by M.Borodin) that `timestamp` is supposed to be in UTC, but
in fact it can be UTC or CET or CEST depending on when and who writes
the data.

Setting timestamp time zone to local time zone meant that those records,
written with proper UTC, are likely to be missed (as we check interval
one or two hours earlier then they are actually written).

Setting timestamp time zone to the UTC means that for records with
timestamp in CE(S)T we won't see changes as soon as they are written to
the Oracle DB; but they will be handled later. It is not best, but still
better than missed records. When things get standardized and all the
update operations start setting UTC timestamp, it will work propertly.