pyDKB: introduce pyDKB.storages submodule.
Sometimes we have to interact with the same storage in different
scripts. When it happens we got to implement same functionality multiple
times: read configuration, check that Python module is available, create
client, ...

This submodule is a place to implement this once and then reuse whenever
it is needed.

Initial module structire:

```
pyDKB
  |---> storages
          |---> exceptions
          |       |---> StorageException
          |       |---> NotFound
          |
          |---> client
                  |---> Client
```