Remove batching of inconsistent records.
Printing all discovered inconsistent records to stdout as a batch contradicts
with various things, such as pyDKB's file mode and the possibility of
controlling the workflow with Apache Kafka.

Create an output message with _id and _type for each inconsistent record.
Still exit with code 1 if at least one inconsistent record was found,
0 otherwise.