pyDKB/storages: make ES client retry request on timeout.
Sometimes we see ConnectionTimeout even with simple `get` request; it
only means that for some reason ES just can't do anything about the
request, not that the request is too heavy.

Now there is a possibility to set number of timeout retries when the
client is created; by default the number is 3.

The ES client itself (`elasticsearch.Elasticsearch()`) by default turned
off the 'retry on timeout' possibilityr, so we have to turn it on 'by
hand'; while the retry number 3 is just the same as default.