DF/009: improve `offset_now()` function.
There was used not very readable way to get current datetime in specific
timezone; new way looks much simplier and gives same result.

Also, it helps in case of undefined `OFFSET_TZ`: if it is `None` and no
`tz` is passed to the function, we will end up with `None.fromutc(...)`,
which will result in `AttributeError`. In the new version same situation
leads simply to `datetime.now(None)`, which is interpreted as "No time
zone specified, so use local time zone." Looks reasonable enough.