DF/019: change default EOM.
As newline symbol (`\n`) is presented in the output message now, we can
not use it as a special symbol EOM (message delimiter). New default
value is ASCII Record Separator symbol (RS, hex: 1E, dec: 30) for both
stream and file modes.

It may seem a better choise to set the EOM to empty line in file mode,
but empty EOM is allowed only if there is one message per file (at least
it is the convention the pyDKB library is going to follow) which is not
true in this case.

When used with Kafka supervisor, this EOM is a matter of communication
between 019 and its supervisor; yet in the shell script
(`data4es-start`), where all stages are linked to each other directly,
via pipes, this EOM must be artifitially added to the stage input
stream, and then removed from its output stream before it gets to the
next stage (which is not capable of handling custom EOM marker yet).

We might want to set custom EOM with `-e` (e.g. `-e '\n'`), but it
does not feel right, because, again, setting EOM to the symbol that
appears in the message body is not a Good Thing. Yet when all stages
will become aware of custom EOM, we may want to set EOM for this process
to RS globally (due to the direct interaction between stages they all
must use same EOM).