Remove DKBFrontEnd from 'master'.
GitHub produces security aletrs, related to NodeJS packages out of date.

May be one day we will decide to restore this functionality and go on
with development/support, but for now it is not used and no one takes
care of keeping it up to date. So I remove it from the repository, but
it will be restored in the new branch 'archive-web-gui-ontodia'.