DF/019: improve service fields handling.
The stage used to remove only specific set of service fields -- those
used by the stage to construct 'action string' for ES bulk api:
'_type', '_id', '_parent' and '_index' (which was not used but would be
removed anyway).

It works as, in fact, the only service fields added in the `data4es`
process are intended to be used for this 'action string' and no other
fields are meant to appear. Still there is some kind of agreement,
saying that any field starting with '_' is a "service" field and must be
ignored when something is ment to be done with data fields only.

In the view of this agreement, now all the fields starting with '_', are
filtered out from input data after performing the service operation that
relies on service fields.