DF/019: allow to reset default ES bulk API action to 'update'.
With `--update` option one can make all the output records prepared for
"update" (instead of "index") operation. It should be useful in case of
shortened/modified dataflow, when some stages are skipped.

As `--update` is a logical (yes/no) option without value, the
parameters handling has slightly changed: we need to `unset()` two
command line arguments only if it is an option with parameter value; if
not [1], then only one.

[1]  `getopt()` sets logical option value to `bool(false)` if
it was specified (not very intuitive, but it is how things work); so
this is what we check.