pyDKB: update version info (0.3.dev).
The logic of version names is silightly changed.
Instead of `X.X-SNAPSHOT` it should be `X.X.dev` now; and when some
piece of development is finished it should be turned into
`X.X.YYYYMMDD`.

Increasing even minor version number supposes some significant changes:
it would be strange to do it for every minor improvement. Yet as we have
such thing as 'version', it should be changed somehow, and the 'release'
date looks fine for this purpose.