DF/025: increase timeout in case of ConnectionTimeout error.
It allows us to execute queries that take up to 240 sec
(30*20^n_retries).
Why not to set the default timeout to this value? Well, I believe we
would like to know if too many requests require more than 30 seconds. If
such thing happens, we might want to do something about the query
itself.