Add a representation of the initial message.
While the operations in the stage are independent - they affect different
parts of the data - they still transform the data before passing it to the
next one. This has no impact on processing of the data, but using the
changed data in error messages can lead to confusion and/or difficulties in
understanding which of the messages contained it. Avoid this by saving the
string representing the initial data and passing it to the functions.

Currently, transform_chain_data() is the only function that displays the data
in error messages. However, this can change in the future, so a quicker
solution - moving transform_chain_data() forward, to be performed first - was
not used.