DF/025: allow execution with missed ES module.
If one does not have Python ES module installed and does not want this
stage to be executed, it is easier to run it with `--skip`, not to
remove it from some existing scenario.

So the stage must not crush even if the ES module is missed; only if it
is missed AND one wants to use it.