DF/091: allow execution with missed/misconfigured rucio module.
If one does not have Python rucio(.client) module installed and/or properly
configured, and does not want this stage to be executed, it is easier to run it
with `--skip`, not to remove it from some existing scenario.

So the stage must not crush even if the module is missed; only if it
is missed AND one wants to use it.

To avoid NameError due to unimported RucioException, in case of import
failure it should be defined to None.