DF/091: fix output datasets processing in `--skip` mode.
One of the functions of this stage is to convert single input message
with task metadata into set of messages with output dataset metadata.

Another one is to set service fields for ES loader.

Although we might want to skip Rucio-related operations, we anyway would
like these operations to be performed; so `skip_process` method was
re-implemented for this case.

---

It means, by the way, that 091 does too many operations. Not that it
should surprise someone, yet it is just another sign of "something
wrong". Will think it over.