DF/095: allow not to initialise client (for `--skip` mode).
To allow `--skip` mode, first of all the stage should allow running
without `--usercert` and `--userkey`: these values are required only
when one is intended to really send some requests to AMI.

To make things alike in the stages, accessing external sources,
functions should not use global variable directly: only access it via
`get_client()` or something like this.

In theory, one can login to AMI in different ways, not only providing
the certificate: e.g., using X509 proxy certificate. So it is not that
useless to try to init the client without cert/key pair as it may seem.

However, the client does not throw any exception if it can not find
proxy certificate: it simply fails to login/password scenario, which is
not supposed in the stage.

So the scenario here is as follows:
* if user provides key/cert pair, try to initialise the client
  in advance (to avoid passing these parameters through all the
  functions to the one that really needs the client);
* if not, leave client uninitialised till it is needed;
* if after the client initialisation conncet mode is 'LOGIN', throw an
  exception (we can't provide login/password manually).