DF/019: add "incomplete" messages handling.
If input message has service field '_incomplete' (set to True/False), it
means that insert and update documents differ.

Case 1 (action: 'insert', _incomplete: true).
  We must keep the information about record incompleteon to update it
  later.

Case 2 (action: 'insert', _incomplete: false).
  Records are "complete" by default, so there is no need to explicitly
  specify this information (though it is allowed).

Case 3 (action: 'update', _incomplete: true).
  On update it is OK to get incomplete input message: that's why it is
  "update", not "insert". Propagating this information to the final
  storage would be incorrect (unless the update sets some fields, known
  before, to NULL) [1]

Case 4 (action: 'update', _incomplete: false).
  If for some reason we have explicit spectification, that the recored
  *is* complete, it worth propagation to the final storage: if
  previously data was *in*complete, now we know for sure that we don't
  need to get some additional information and update this record later.

In case of 'update' action we have possibility to specify that there is
no difference in data for actual update and for "upsert" (insert in case
of missed record in the storage). It allows us to transfer less data
(one document instead of two), thus worth using. But it works only if we
know nothing about input message (in)completeness (service field
'_incomplete' is not specified), or know for sure that it is complete
('_incomplete' is set to False). In the latter case, to avoid passing
two documents (one with 'update_required' set to False, another without
it), it looks fine to insert document with this (not required) field.

---

[1] So we must ensure that all the stages leave undefined fields
    empty, not setting them to NULL or empty string or whatever else,
    that can erase known values in the final storage.