Fix the interruption of a cycle.
Per Python documentation, "(continue statement) continues with the next cycle
of the nearest enclosing loop". Therefore, "continue" makes no sense as
the last line in the loop. It seems that "continue" was mistakenly used in
place of "break": the goal of the inner loop is to locate two "fields", one
with the parameter's name and the other with its value. For example, the
processing of the following fields:

    ...
    {
      "@table": "dataset_physics_bridge",
      "@name": "insert_time",
      "@type": "TIMESTAMP(11)",
      "$": "2016-10-26 01:16:59"
    },
    {
      "@table": "physicsparameterdefs",
      "@name": "paramName",
      "@type": "VARCHAR2(128)",
      "$": "genFiltEff"
    },
    {
      "@table": "dataset_physics_bridge",
      "@name": "paramValue",
      "@type": "VARCHAR2(128)",
      "$": "1.0000E+00"
    },
    ...

will result in the assignment:

    data["genFiltEff"] = "1.0000E+00"

Due to this and the assumption that each "field" can only be used once in each
row, replace "continue" with "break": why check the remaining fields when the
two necessary are already located?
Remove the line setting variables to None: with "break" the outer loop will
proceed to next cycle and they will be set to None there.