Rework container name determination.
Operation of retrieving the container name is not related to query to AMI.
Currently, it is split in two parts, the first (retrieval of dataset name
from dataset info) being performed in amiPhysValues() and the second (removal
of '_tid...' from dataset name) in remove_tid(). It would be logical to
join the two parts together outside of amiPhysValues().

- Rename remove_tid() to container_name().
- Move dataset name extraction from amiPhysValues() to container_name().
- Update the description of container_name() to adapt to changes.