ES/mapping: change field type for "pr_id" (keyword -> integer).
Keyword type is more useful for string, not numeric values: it allows
wildcard search, but not range search. Range query looks more natural,
so base type for this field is changed to numeric.

There are production request ID values >= 20000, so 'short' looks a bit
risky. But it does not look like the value can exceed 2^31-1.

However for "google-like" search wildcard might still be useful, so
keyword field is stored as well.