ES/mapping: change field type for "energy_gev" (long -> integer).
Short allows values up to 32767, which gives about 32TeV.
As far as I know, such CME is not planned for LHC, so it would be good
enough... if there were no mistakes in Production Request definitions /
descriptions in ProdSys database: some PRs claim to have energy 140000
or 130000 GeV (namely, pr_id: 1607, 4785, 3871). It is obviously a typo,
but for now I don't think that DKB must somehow fix it during data
integration: who can tell for sure, how many different types of typos we
should be able to handle?

So `integer` it is.