ES/mapping: change field type for "run_number" (keyword -> integer).
Keyword is better to be used for string values, not numeric: for numbers
its only effect is that we can perform wildcard serach -- but not range
search. But range search looks more natural, so...

Current values of run number are about 0.5M, so integer (up to 2^31-1)
looks good enough.

However, "keyword" (google-like) search might still require wildcard
search, so "keyword" field is to be stored.