ES/mapping: add service (marker) field "update_required".
If this field is set to True, it means that during the metadata
integration process some stages were skipped or failed. This marker is
to be used to re-process "incomplete" documents and get the "complete"
version (and do it only for those documents that *need* the update, not
for the whole archive).