API/setup: fix issue with socket permissions.
Socket should be read/writeable for nginx user, which can not be granted
by any user. Still there may be a special directory for sockets in the
system, from which the socket file will inherit group permissions.

When installed with root privileges, this directory will be created and
properly configured. When setup script run by non-root user, which also
is not in nginx group, only already created directory can be used. For
this purpose the directory is configured with full access for any user
-- while files created in the directory during installation and when the
application is started are accessible only for the owner and nginx group
(so one application will not be able to interfire with another one by
messing with its files).