API/setup: improve `--cfg-dir` handling.
In previous version if WWW dir was changed by cmdline parameter
`--dest`, in config directory default path still was used original
(default) value of `$WWW_DIR`, so if `--dest` is used `--cfg-dir` must
be used as well.

Now value of `$CFG_DIR` (default or passed via `--cfg-dir`) is treated
as-is if it is an absolute path (starts with "/" or, in case of
`--cfg-dir`, with "." to point to a file in a local (sub)directory);
else it is considered as a subdirectory of `$WWW_DIR` (whatever value it
has).