API/kwsearch: fix search by specially analyzed fields (e.g. taskname).
By default ES "query string" searches in special "_all" pseudo-field:
its content is combined of all document fields content, separated by
spaces. And this "_all" field is analyzed as an ordinary text field,
which means that DS/task names are not tokenized by dots and
underscores. So if DS/task name contains "phys_short" field that looks
like "aaa_bbb_ccc", search by "bbb" will return nothing (unless there's
"bbb" mentioned in other fields, such as "description").

If "all_fields" parameter of "query string" is set to True, ES will
search for the passed search string in all fields mentioned in the
mapping, taking them tokenized according to the mapping (and tokenizing
the expressions in the query string for each field according to the
mapping).

ES UPGRADE INFO: "all_fields" was removed in 7.0 (in favour of
"default_field": "*"). But it was done together with some fixes that
would prevent "*" from being interpreted as all -- both data ans service
-- fields; so for now (having ES v.5.4.1) we can not use this variant.
But whin we decide to upgrade -- it should be fixed.