API/setup: make sure that install script uses proper version of Python.
On `bamboo` we have an old system which works with Python 2.6.
When install script is run with `scl enable python27`, it uses Python
2.7 to run API application. But to compile .py files from proper user it
runs `sudo -u <user> python ...`, and here user's environment is used,
so we fall back to Python 2.6.

To use Python 2.7 it needs to know where to look for library files, so
`LD_LIBRARY_PATH` is passed to the execution environment as well.

`python_exec` is specifically set to 2.7 to make sure that if someone
runs the script from with the Python 2.6, it will produce an error.