API/setup: fix issue with incorrect group and permissions of SOCK_DIR.
Sometimes installation script fails to create SOCK_DIR with proper
group ownership and permissions. It happens when:
* `/var/run` is a symlink to `/run` (as in CentOS 7);
* installation script is run with explicitly specified RUN_DIR
  (`-R /var/run/wsgi`) and default (hardcoded) SOCK_DIR
  (`/var/run/wsgi`).

In this case RUN_DIR is actually set to `/run/wsgi` (due to `readlink`).
Then in `ensure_dir` it is managed before SOCK_DIR and, if desired
directory is missed, created as RUN_DIR with no specific settings
applied (since `/run/wsgi == /var/run/wsgi` is false).
And when we get to SOCK_DIR, it already exists and will be skipped --
and these specific settngs are not applied again.

Now we directories themselves, not strings -- and it works correctly.