Add function to get output formats by amitags.
Original code (in Big PanDA Monitoring) uses Oracle table T_STEP_TEMPLATE
to construct a list of output formats by given ami tags, and uses it in
calculation of derivation statistics. Implement a function for such
construction that, for now, uses a json file instead of Oracle. Attach
the function to (currently) not implemented derivation handler.