ES/backup: use ES 'snapshot' mechanism to avoid archiving "live" directory.
If we work with "living" instance of ES, it may do something with data
at the moment we are trying to archive them -- and in this case we will
have an inconsistent copy of the data directory, which is useless to
restore indexed data.

Current version is applicable only to the standalone (single-node) ES
instance, since in cluster mode every node will create a snapshot in its
local directory. To work in this mode, we may need to split the task in
two: snapshot update and snapshot backup -- and make sure that the
second part won't start before the first one is completed -- or configure
'remote repository' to store whole snapshot on a single node.