Rework format retrieval by means of aggregation.
ES was designed with handling big amounts of data in mind, so it is
logical to use its capabilities where it is prudent.
Current version of code is executing query "get data formats of all output
datasets with parents having given ctags" and then manually
transforming the resulting set of lists into a list of unique occurences.
Replace it with execution of aggregation query "get all data formats of
datasets with parents having given ctags", where uniqueness is guaranteed
by aggregation mechanisms.