pyDKB/stage: add possibility to (re)set default values for arguments.
It is useful when a custom stage implementation knows something about
arguments that is not applicable to any stage (e.g. default config file
location).

I didn't find a way to make argument "required", and argparse
documentation says:
{{{
Required options are generally considered bad form because users expect
options to be optional, and thus they should be avoided when possible.
}}}

So I decided that it would be good enough to allow specifying default
values.