DF/095: replace `--usercert` and `--userkey` with `--config`.
It is more convenient to provide configuration as a file, not as command
line parameters (which, actually, was already done via `data4es-start`);
and it is better if a stage can read its configuration itself, without
some wrapper that transforms found configuration parameters into a
command line arguments.

Now instead of two command line arguments (`--usercert` and `--userkey`)
one should provide path to a configuration file that contains values of
these two parameters in the following form:
```
[stage]
usercert=/path/to/crt
userkey=/path/to/key
```

The default value for config file is set to `../config/095.cfg`,
relative to the stage directory (`Utils/Dataflow/config/095.cfg`).