DF/025: allow CPU usage calculation for all tasks.
Previously it was done only for the finished ones -- it was the part
of the task description provided by M.Borodin (since originally it was
supposed that we take this information from BigPanDA that would not
be happy if we queried it every now and again).

Now we're gonna get this information for all tasks: whatever's its
state, if there are some jobs -- we can get some useful data.