DF/025: replace `task_data` doc_type with default (`_doc`).
Due to the change in the Chicago ES after upgrade, doc types are no
longer supposed to be used.
Although `tasks_archive_2020` index is already created and doc_type of
new documents is still `task_data`, one day it will become `_doc` and I
don't want us to stumble over it again.

When doc_type is not specified, ES Search API searches all documents in
the index independently of their type, and it doesn't seem like there
are any other documents but those we use.