DF/091: improve error handling ('_incomlete' messages).
Changes in error handling.
1. If Rucio returns `DatasetIdentifierNotFound`, set `deleted` to
   `True` (previously was treated as any other `RucioException`).
2. In case of `RucioException` mark message as "incomplete" -- this
   record will be written to ES in "update" mode, not "insert"
   (previously `deleted` and `bytes` would be set to `True` and `-1`).
3. If some of required fields can not be extracted from Rucio (or have
   value `null`), they are left unset (previously would be set to
   `None`.
4. If in the result message some fields of those that were supposed to
   be added at this stage are missed, the message is marked as
   `_incomplete`.

What is not good in this logic: if dataset was removed from Rucio, it
will always be marked as "requiring update". But for now we have only
this option: to use "update" instead of "insert", we need to know that
the message is incomplete; and when the message coming to Stage 019 is
incomplete -- it should be marked as "requiring update" for further
investigation.

Maybe the logic could be extended and "_incomplete" should be turned
into two different markers: "update since the original source removed
useful data" and "update since we failed to connect to the original
source", but... not now.