API/campaign_stat: add alternative version for output events calculation.
Affected fields are:
* data['overall_events_processing_summary'][<step>]['output'];
* data['events_24h'][<step>].

Previously for these fields was used total number of events in all
output datasets. But some datasets might be removed (e.g. pre-merge
datasets). It may be not actual for the most recent data, but still...

Now for this value can be used sum of `processed_events` fields of tasks in
`finished` or `done` state. The `processed_events` field is calculated as:
* sum(ATLAS_PANDA.JEDI_DATASETS.neventsused) for 'input' datasets with
  NULL 'masterid';
* ATLAS_DEFT.T_PRODUCTION_TASK.total_events (if there's no such
  datasets).

To switch between these versions added request parameter `events_src`
that acsepts values:
* 'ds'   -- use output datasets version;
* 'task' -- use `processed_events`;
* 'all'  -- provide all possible values as hash.

If 'all' is specified, field
`data['overall_events_processing_summary'][<step>]['ratio']` will be set
to `null`.