DF/069: add verification for the input data.
If input data are broken it'd be nice to know about it. However the ES
bulk API seems to not not throw away 'action' strings with, e.g., last
curly bracket missed, and successfully load such records. So we also do
not throw away "incorrect" records and just log their presence.