API/task_stat: ES query implementation (pt.1): "query" part.
Currently query only returns number of `task` documents matching
following conditions:
* field conditions passed by user (e.g. `?htag=newfastcalosimntuponly`
  or `?pr=18030`);
* task status is not one of: 'aborted', 'failed', 'broken', 'obsolete'.

Filtering of "bad" tasks is added after the ProdTask implementation of
this functionality:
https://github.com/PanDAWMS/panda-bigmon-atlas/blob/1e0dcdb0c8fa590f697311ba4a422a7c273b2269/atlas/dkb/views.py#L385
https://github.com/PanDAWMS/panda-bigmon-atlas/blob/1e0dcdb0c8fa590f697311ba4a422a7c273b2269/atlas/dkb/views.py#L637
https://github.com/PanDAWMS/panda-bigmon-atlas/blob/1e0dcdb0c8fa590f697311ba4a422a7c273b2269/atlas/dkb/views.py#L755

My guess is that for these "bad" tasks there are "good" ones, so having
both "bad" and "good" ones would double their values for input/requested
events and input data volume, which is not a good thing if we want these
values to represent "how many data are (being) processed" and "how many
events should be processed in the end".

Although it *might* be useful to get "bad" values for some purposes, for
now it is not needed.

---

The query and its results are written to the server application log for
debugging/verification purposes.