API/task_stat: ES query implementation (pt.2): steps aggregation.
Steps aggregation is the highest level of the aggregation that divide
all selected tasks into number of buckets that answer processing steps.
All other aggregations will be performed within these buckets == for
every step.

Current version of the query only returns total number of documents
matching the selection query and number of documents for each step.

--

The query and its results are written to the server application log for
debugging/verification purposes.