pyDKB/stage: allow resetting custom default values to 'safe' ones on skip.
If stage is skipped, some things that are supposed to be properly
configured for real processing, may not be configured.

In theory, it should be possible to run stage with all command line
arguments having their default values (er, except for the `--mode`
parameter, which has default value 'f' for file processing -- which
means that if no file specified, stage will fail). However for each
stage there may be additional restrictions -- like config file, etc.
required for the real processing. So one may decide to specify some
default values in the code wherever it is possible.

But in case of "skip" scenario the real processing won't happen, so
it may not be critical to have correct values: e.g. the "default"
configuration file with authentication parameters for some external
source may be missed, but since we're not going to actually connect
there -- it can be ignored.