DF/091: update samples.
`/Utils/Dataflow/test/utils/compare_ndjson_files.py` shows changes like these:

```
Record seem to differ for uid=mc15_13TeV.387007.MGPy8EG_A14N_BB_direct_300_295_MET100.merge.AOD.e3994_a766_a821_r7676_tid08124406_00
key = deleted:
Items missed in (2):
(1) False
Items missed in (1):
(2) True
Item missed in (2): 'bytes'
Item missed in (2): 'events'
Key missed in (1): '_incomplete'
```

...which means:
* DS with given uid was removed ('deleted: True');
* fields 'bytes' and 'events' are not presented in the new record;
* key '_incomplete' is added.

So when this record gets to the 019 (esFormat) and 069 (upload2es), it
will be written in the "update" mode: 'bytes' and 'events' won't be
changed, 'removed' will be set to True, and '_update_required' will be
set to True.

The latter may be not the best option, but it is the only flag we have
right now to say that there's something wrong about given record... and
here "wrong" is that Rucio did not provide us with the information we
needed.