pyDKB/df/messages: improve the 'incomplete' message marker handling.
When a (JSON) message is marked as 'incomplete', it used to mean that
its 'decoded' version is adjusted. Which is not good: 'incompletness',
in fact, has nothing to do with the message content, so no matter if it
is complete or not -- the content is a content and must not carry any
service information.

Now the 'incompleteness' is to be stored strictly in the instance
variable and become 'visible' only in the encoded version of the message
(so that the parameter could be restored in the next stage). For this
purpose, if message is initialized with 'encoded' data, it is to be
decoded on the initialisation (to get correct 'incompleteness' status).

It also means that 'encoded' version:
* must be produced not from the original data, but form the "content"
  (for original data, if provided in 'encoded' form, may contain
   an unwanted incompleteness marker);
* must have the cached value reset whenever the incompleteness status
  has changed.

NOTE: if a (JSON) message is initialized with data provided in a
'native' format, the 'incompleteness' status won't be extracted from the
data and will be treated simply as another data field.