pyDKB/df/messages: remove "naive" decoding form message __init__().
"Naive" decoding is a "decoding" based on the input data format: if it
is one of the native ones, passed data are taken as "already decoded".

However even the native format may contain some service information
(like 'incompleteness' marker), and it's better to know about it. Else
it may be confusing: e.g. for JSON message initialized with data
containing '_incomplete: True' the 'incomplete()' method would return
False (which is strange if one knows how the information about messages
incompleteness is passed between stages).