pyDKB/test: fix issues related to pyDKB line numbers changes.
When pyDKB code is updated, some test cases that contain traceback info
in the expected STDERR, can be broken. From now on all ', line XXX' parts
of these tracebacks in the `case/*/err` can (and should) be written as
', line <NNN>' -- this way, the real line numbers are hidden and their
changes won't affect the test cases.