API/campaign_stat: change 'events_daily' statistics calculation.
"Daily progress" is not "last day's progress", but "day-by-day
progress". So the method's output data format has changed: instead of
'events_24h' there's now 'events_daily_progress' sections, with data in
the following form:

```
{
  ...
  "data": {
    ...
    "events_daily_progress": {
      <step_name>: {
        <date>: <n_events_processed_by_tasks_finished_'today'>,
        ...
      },
      ...
    }
  }
}
```

Value for `<n_events_processed_by_tasks_finished_'today'>` is calculated
as sum of output events of all tasks with "end_date" within given day's
interval (from <date> 00:00:00 to <date> 23:59:59) and with status
'done' or 'finised'.

Output events, as in other cases for this method, can be calculated
basing on the task's "processed_events" field, or on the number of
events in the task's (not removed) output datasets (see
`?events_src=...` method's parameter).

---

Example of the method's response:

```
{
  "status": "OK",
  "took_total_ms": 960,
  "total": 12682,
  "data": {
    "events_daily_progress": {
      "NTUP:p3474": {
        "08-03-18 00:00:00": 11950000.0,
        "07-03-18 00:00:00": 2430000.0,
        "10-03-18 00:00:00": 150000.0,
        "12-03-18 00:00:00": 80000.0,
        "13-03-18 00:00:00": 10000.0,
        "11-03-18 00:00:00": 90000.0,
        "09-03-18 00:00:00": 950000.0
      },
      "NTUP:r10284": null,
      "NTUP_FCS:p3449": {
        "26-02-18 00:00:00": 618000.0,
        "01-03-18 00:00:00": 68000.0,
        "25-02-18 00:00:00": 570000.0,
        "23-02-18 00:00:00": 1010000.0,
        "24-02-18 00:00:00": 550000.0,
        "28-02-18 00:00:00": 60000.0,
        "22-02-18 00:00:00": 250000.0,
        "27-02-18 00:00:00": 110000.0
      },
      ...
    },
    ...
  }
}
```