DF/091: fix issue with undefined `DataIdentifierNotFound`.
When the stage fails to load `rucio.client` module, it does not stop the
execution -- for in case of `--skip` opotion specified we do not
actually need the module to go on.

The script exits only when tries to initialize the client -- but since
the initialization occures within `try/except` clause, it first stumbles
over the expected exception name (`DataIdentifierNotFound`), which it
failed to load from the `rucio.client`. We had similar issue with
`RucioException` once, so now the new exception is handled similarly.