pyDKB/stage: fix error logging in case of `SystemExit`.
`SystemExit` exception, when casted to `str`, gives string with exit
code; so it is logged as `(ERROR) 1`.

So instead of logging any exception's message as ERROR, we now do it
only when we do not have any specific rules about this kind of
exceptions.

It may affect situation when the stage was interrupted with something
like `raise KeyboardInterrupt("User interrupted the program with
Alt+J")` -- it won't log the user's message. But for now we do not have
examples of this usage of `KeyboardInterrupt`, so nothing is broken.