DF/091: improve client initialisation errors handling.
Rucio client module may raise not only `RucioException`, but also some
other exceptions: e.g. `IOError` (when files provided in the
configuration as user certificate/key can not be read).

What we need to do in case of a error in `init_rucio_client()` is to
make sure that this error can be distinguished from any other error
occured during the rucio client usage. In other words -- when we fail to
initialise the client, it is not a "Rucio error" (indicating that we
have some problems with Rucio), but a "dataflow error" (indicating that
we have some problems with the stage which is unable to do its job).

Calling `sys.exit()` is a bit severe: yes, we need to stop the stage
execution if we can't initialise the client; but it would also be
correct to say that we need to stop the dataflow. And, if
`DataflowException` is raised during the `process()` execution, it is
up to the common library to take care about the problem (output
necessary details and interrupt the stage -- or do whatever is the
default action when a stage says "I have a problem operating as a part
of the dataflow process").