DF/069: ensure correct exit code.
No matter what's the last action in the `load_*` function, if we get to
its end (not exited somewhere in between) -- then everything's fine and
the stage as a whole should return 0.

Without this extra "return 0" in the end function returns the last
return code -- which in case of the `load_files` was the return code of
`[ ${PIPESTATUS[0]} -ne 0 ]`. And if the status *was* 0, the return
value of the test was 1.

I could replace `[ . -ne . ] && exit .` with `[ . -eq . ] || exit .`,
but next time we'd stumble over something else, for sure.