API/campaign_stat: improve 'input_events' calculation.
We must make sure that we calculate the total number of input events
only for "valid" tasks (those that were not cancelled, failed, broken or
aborted for some reason). Invalid tasks are very likely to be replaced
with the valid ones, and we don't want to take the number of input
events from both of them.